continued restructure
